refactor(playground): close architecture review findings
