+++ playground/package.json
@@ -20,6 +20,7 @@
     "verify:dist": "node scripts/verify-dist-wasm.mjs",
     "verify:opaque-realm": "node scripts/verify-opaque-realm.mjs",
     "verify:wasm-inputs": "node ../platforms/web/scripts/verify-wasm-inputs.mjs --package full",
+    "verify:editor-artifact-authority": "node scripts/editor-artifact-measurement/verify-editor-artifact-receipt.mjs",
     "measure:editor-artifacts": "node scripts/editor-artifact-measurement/measure.mjs",
     "generate:examples": "cargo run -p xtask -- gen-playground-example-catalog",
     "verify:examples": "cargo run -p xtask -- verify-playground-example-catalog",
@@ -29,10 +30,12 @@
     "test:zenuml-browser-admission": "node scripts/zenuml-browser-admission.mjs --run",
     "record:zenuml-browser-admission": "node scripts/zenuml-browser-admission.mjs --run --write",
     "verify:dependencies": "npm ls --all && npm --prefix tests ls --all",
+    "verify:security": "node scripts/verify-npm-audits.mjs",
     "test": "npm run prepare:browser-runtime && npm run test:prepared",
-    "test:prepared": "npm run test:browser:typecheck && npm run test:build-graph && npm run test:csp-policy:prepared && npm run test:editor-worker && npm run test:mermaid-config && npm run test:runtime && npm run test:mermaid-requirements && npm run test:realm && npm run test:examples && npm run test:export && npm run test:benchmark",
+    "test:prepared": "npm run test:browser:typecheck && npm run test:build-graph && npm run test:dependency-security && npm run test:editor-artifact-decision && npm run verify:editor-artifact-authority && npm run test:csp-policy:prepared && npm run test:editor-worker && npm run test:mermaid-config && npm run test:runtime && npm run test:mermaid-requirements && npm run test:realm && npm run test:examples && npm run test:export && npm run test:benchmark",
     "test:benchmark": "node --test --experimental-strip-types src/benchmark/*.test.ts src/benchmark/realm/*.test.ts",
     "test:build-graph": "node --test scripts/typescript-source-graph.test.mjs scripts/vite-manifest-graph.test.mjs scripts/playground-build-policy.test.mjs scripts/runtime-module-request-policy.test.mjs scripts/opaque-realm-artifact-plan.test.mjs scripts/opaque-realm-artifact-verifier.test.mjs scripts/validation-migration-ledger.test.mjs scripts/benchmark-page-guards.test.mjs scripts/csp-policy.test.mjs scripts/package-entrypoints.test.mjs",
+    "test:dependency-security": "node --test scripts/verify-npm-audits.test.mjs",
     "test:csp-policy:prepared": "node --test scripts/csp-policy.prepared.test.mjs",
     "test:examples": "node --test --experimental-strip-types src/lib/examples.test.ts",
     "test:editor-worker": "node --test --experimental-strip-types src/editor/protocol.test.ts src/editor/worker-client.test.ts src/editor/worker-runtime.test.ts src/lib/mermaid-language.test.ts",
