+++ playground/src/editor/protocol-version.ts
@@ -1,0 +1,2 @@
+export const EDITOR_WORKER_PROTOCOL = 3 as const;
+export const EDITOR_SCHEMA_VERSION = 1 as const;
