+++ tools/upstreams/ZENUML_CORE_CANDIDATE_EVIDENCE.json
@@ -2399,7 +2399,7 @@
     "probeCount": 13,
     "observationCount": 13,
     "passedObservationCount": 13,
-    "artifactSha256": "5687e9b1d662b4400ff4e401177eb1fc4f164e54bffe25d6c43fed92d122aa57"
+    "artifactSha256": "63b473c2bbcbe768397cff567ae1564b9457a74f2b4523cf8815cea90e1d3e21"
   },
   "security": {
     "artifact": "tools/upstreams/ZENUML_BROWSER_SECURITY_EVIDENCE.json",
@@ -2408,7 +2408,7 @@
     "probeCount": 18,
     "observationCount": 18,
     "passedObservationCount": 18,
-    "artifactSha256": "5687e9b1d662b4400ff4e401177eb1fc4f164e54bffe25d6c43fed92d122aa57"
+    "artifactSha256": "63b473c2bbcbe768397cff567ae1564b9457a74f2b4523cf8815cea90e1d3e21"
   },
   "resource": {
     "measurementScope": "runtime-entry",
